+++ src/commands/controllers/files/guard.ts
@@ -1,0 +1,124 @@
+import {
+  getClass,
+  getFolder,
+  parsePath,
+  save,
+  toKebabCase,
+} from '../../utils/functions';
+
+const content = `import { Injectable } from '@angular/core';
+import {
+  ActivatedRouteSnapshot,
+  CanActivate,
+  CanActivateChild,
+  CanDeactivate,
+  CanLoad,
+  CanMatch,
+  Route,
+  RouterStateSnapshot,
+  UrlSegment,
+  UrlTree,
+} from '@angular/router';
+import { Observable } from 'rxjs';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class {className}Guard
+  implements
+    CanActivate,
+    CanActivateChild,
+    CanDeactivate<unknown>,
+    CanLoad,
+    CanMatch
+{
+  canActivate(
+    route: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    return true;
+  }
+  canActivateChild(
+    childRoute: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    return true;
+  }
+  canDeactivate(
+    component: unknown,
+    currentRoute: ActivatedRouteSnapshot,
+    currentState: RouterStateSnapshot,
+    nextState?: RouterStateSnapshot
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    return true;
+  }
+  canMatch(
+    route: Route,
+    segments: UrlSegment[]
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    return true;
+  }
+  canLoad(
+    route: Route,
+    segments: UrlSegment[]
+  ):
+    | Observable<boolean | UrlTree>
+    | Promise<boolean | UrlTree>
+    | boolean
+    | UrlTree {
+    return true;
+  }
+}
+`;
+
+const newGuard = async (vscode: any, fs: any, path: any, args: any = null) => {
+  let relativePath = '';
+
+  if (args) {
+    relativePath = parsePath(vscode, path, args);
+  }
+
+  const folder = await getFolder(
+    vscode,
+    'Folder name',
+    'Folder name. E.g. src, app...',
+    relativePath,
+  );
+
+  let className = await getClass(
+    vscode,
+    'Guard class name',
+    'E.g. User, Role, Auth...',
+  );
+
+  if (className === 'Guard') {
+    vscode.window.showErrorMessage('The file has not been created!');
+    return;
+  }
+
+  className = className.replace(/guard/gi, '');
+
+  const body = content.replace(/\{className\}/g, className);
+
+  const filename = '/' + folder + toKebabCase(className) + '.guard.ts';
+
+  save(vscode, fs, path, filename, body);
+};
+
+export { newGuard };