+++ src/commands/controllers/files/interceptor.ts
@@ -1,0 +1,68 @@
+import {
+  getClass,
+  getFolder,
+  parsePath,
+  save,
+  toKebabCase,
+} from '../../utils/functions';
+
+const content = `import { Injectable } from '@angular/core';
+import {
+  HttpRequest,
+  HttpHandler,
+  HttpEvent,
+  HttpInterceptor,
+} from '@angular/common/http';
+import { Observable } from 'rxjs';
+
+@Injectable()
+export class {className}Interceptor implements HttpInterceptor {
+  intercept(
+    request: HttpRequest<unknown>,
+    next: HttpHandler
+  ): Observable<HttpEvent<unknown>> {
+    return next.handle(request);
+  }
+}
+`;
+
+const newInterceptor = async (
+  vscode: any,
+  fs: any,
+  path: any,
+  args: any = null,
+) => {
+  let relativePath = '';
+
+  if (args) {
+    relativePath = parsePath(vscode, path, args);
+  }
+
+  const folder = await getFolder(
+    vscode,
+    'Folder name',
+    'Folder name. E.g. src, app...',
+    relativePath,
+  );
+
+  let className = await getClass(
+    vscode,
+    'Interceptor class name',
+    'E.g. User, Role, Auth...',
+  );
+
+  if (className === 'Interceptor') {
+    vscode.window.showErrorMessage('The file has not been created!');
+    return;
+  }
+
+  className = className.replace(/interceptor/gi, '');
+
+  const body = content.replace(/\{className\}/g, className);
+
+  const filename = '/' + folder + toKebabCase(className) + '.interceptor.ts';
+
+  save(vscode, fs, path, filename, body);
+};
+
+export { newInterceptor };