+++ src/commands/controllers/terminal/environments.ts
@@ -1,0 +1,7 @@
+import { execute } from '../../utils/functions';
+
+const generateEnvironments = (vscode: any) => {
+  execute(vscode, 'generate environments', 'ng generate environments');
+};
+
+export { generateEnvironments };