+++ CHANGELOG.md
@@ -7,9 +7,43 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2023-12-08
+
+### Added
+
+- Add file generetor functions
+- Add Disable Analytics command
+- Add Enable Analytics command
+- Add Info Analytics command
+- Add Prompt Analytics command
+- Add Clean Cache command
+- Add Disable Cache command
+- Add Enable Cache command
+- Add Info Cache command
+- Add Generate Component command
+- Add Generate Environments command
+- Add Start Server command
+- Add export class Module snippet
+- Add export class RoutingModule snippet
+- Add export class Component snippet
+- Add export class Service snippet
+- Add unsubscribe snippets
+- Add ngOnDestroy snippet
+- Add ngOnInit snippet
+- Add export const environmentsnippet
+- Add @if, @else snippets
+- Add @for snippet
+- Add @empty snippet
+- Add @switch snippet
+- Add @case snippet
+- Add @default snippet
+- Add router-outlet snippet
+- Add router-link snippet
+
 ## [0.1.0] - 2023-12-08
 
 - Initial release
 
-[unreleased]: https://github.com/ManuelGil/vscode-angular-generator/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/ManuelGil/vscode-angular-generator/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/ManuelGil/vscode-angular-generator/compare/v0.1.0...v1.0.0
 [0.1.0]: https://github.com/ManuelGil/vscode-angular-generator/releases/tag/v0.1.0